refactor(dekstop): move commands to separate module and refactor mod conflict dialog
