+++ apps/desktop/src/components/mod-browsing/hero-conflict-dialog.tsx
@@ -29,8 +29,7 @@ interface HeroConflictDialogProps {
 }
 
 const prefersReducedMotion = () =>
-  typeof window !== "undefined" &&
-  window.matchMedia?.("(prefers-reduced-motion: reduce)").matches;
+  window.matchMedia("(prefers-reduced-motion: reduce)").matches;
 
 export const HeroConflictDialog = ({
   open,
