+++ apps/desktop/src/components/mod-browsing/mod-button.tsx
@@ -1,8 +1,3 @@
-import { DeadlockHeroes } from "@deadlock-mods/shared";
-
-function isDeadlockHeroKey(key: string): key is keyof typeof DeadlockHeroes {
-  return key in DeadlockHeroes;
-}
 import type { ModDto } from "@deadlock-mods/shared";
 import { Button } from "@deadlock-mods/ui/components/button";
 import { toast } from "@deadlock-mods/ui/components/sonner";
@@ -244,11 +239,8 @@ const ModButton = ({ remoteMod, variant = "default" }: ModButtonProps) => {
                 m.detectedHero === detectedHero,
             );
             if (conflictingMod) {
-              const heroDisplay = isDeadlockHeroKey(detectedHero)
-                ? DeadlockHeroes[detectedHero]
-                : detectedHero;
               const resolution = await askHeroConflict(
-                heroDisplay,
+                detectedHero,
                 conflictingMod,
                 localMod,
               );
