+++ apps/desktop/src/hooks/use-hero.ts
@@ -4,7 +4,7 @@ import { getHeroByName } from "@/lib/deadlock-api";
 export const useHero = (name: string | null | undefined) =>
   useQuery({
     queryKey: ["deadlock-hero", name],
-    queryFn: () => getHeroByName(name as string),
+    queryFn: () => getHeroByName(name!),
     enabled: !!name,
     staleTime: 1000 * 60 * 60,
     gcTime: 1000 * 60 * 60 * 24,
